Restore current page after user-initiated login
With this change, login should now feel seamless - either the user returns to where they were (for manual logins) or goes to where they wanted to go (for auto-logins on actions requiring authentication).
