+++ lib/db/StudyDAO.js
@@ -41,7 +41,8 @@ SELECT
   WHERE sr."centrelineId" = $(centrelineId)
   AND sr."centrelineType" = $(centrelineType)
   AND sr."status" != 'COMPLETED'
-  AND NOT sr."closed"`;
+  AND NOT sr."closed"
+  ORDER BY s."createdAt" DESC`;
     return db.manyOrNone(sql, { centrelineId, centrelineType });
   }
